+++ src/dipdup/cli.py
@@ -426,8 +426,6 @@ async def hasura(ctx) -> None:
 async def hasura_configure(ctx, force: bool) -> None:
     """Configure Hasura GraphQL Engine to use with DipDup."""
     config: DipDupConfig = ctx.obj.config
-    url = config.database.connection_string
-    models = f'{config.package}.models'
     if not config.hasura:
         raise ConfigurationError('`hasura` config section is empty')
     hasura_gateway = HasuraGateway(
@@ -436,9 +434,12 @@ async def hasura_configure(ctx, force: bool) -> None:
         database_config=cast(PostgresDatabaseConfig, config.database),
     )
 
-    async with tortoise_wrapper(url, models):
-        async with hasura_gateway:
-            await hasura_gateway.configure(force)
+    async with AsyncExitStack() as stack:
+        # TODO: Verify and prepare models in tortoise_wrapper?
+        await DipDup(config)._set_up_database(stack)
+
+        await stack.enter_async_context(hasura_gateway)
+        await hasura_gateway.configure(force)
 
 
 @cli.group()